move parse data
