+++ src/views/Transactions/Transactions.tsx
@@ -2,10 +2,11 @@ import { Wrapper, Title, ConnectButton, Account } from "./Transactions.styles";
 import useTransactionsView from "./useTransactionsView";
 import TransactionsTable from "./TransactionsTable";
 import { shortenAddress } from "utils/format";
+import formatTransactions from "./formatTransactions";
 
 const Transactions = () => {
-  const { isConnected, initOnboard, account, rows, headerCells } =
-    useTransactionsView();
+  const { isConnected, initOnboard, account } = useTransactionsView();
+  const { rows, headers } = formatTransactions();
 
   return (
     <Wrapper>
@@ -20,7 +21,7 @@ const Transactions = () => {
       )}
       {isConnected && (
         <>
-          <TransactionsTable rows={rows} headerCells={headerCells} />
+          <TransactionsTable rows={rows} headers={headers} />
         </>
       )}
     </Wrapper>
